2025_q1/requirements.txt
google-generativeai
google-genai
rich